Returning correct songs added to file
When printing the number of songs added to the file, it returns the total number of songs found without subtracting the number of songs skipped. Initialized a zeroed variable that will count as songs are skipped and subtract from the total songs found.  Not 100% positive this is the best/correct way, just my best guess.